docsis.py: Use tempfile to generate temp directory name
I was getting a problem trying to add some simple code to boardfarm
and the dependancy here on 'config' was one issue.

The 'tempfile' libary is standard in python for generating temporary
file or directory names. Using this instead of station name from a
config reduces dependancies and makes managing this code easier.

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: I724208466681b26cfea6cb659adeecaa7f2b6e1c
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm-docsis/+/370
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Matthew McClintock <matthew@meldnetworks.com>