tests/docsis_lib/docsis: Slightly different logic for loop over mibs_paths
This is the first step towards eliminating the 'from devices import board'.

If a mibs_path is passed in, use that. If no mibs_path passed in, use
board.mibs_path. If that doesn't even exist it will be an empty list.

To Do: Change every call of docsis() so that it passes in board.mibs_paths.
Currently no call passes in any mibs_paths. Once all calls of docsis pass
in mibs_paths, I will delete the 3 lines of code:

    from devices import board
    if mibs_paths == []:
        mibs_paths = getattr(board, 'mibs_paths', [])

Signed-off-by: Mike Anderson <mbanderson@uwalumni.com>
Change-Id: Id4df494a8eb5a73283fb0897e8e978af145b3076
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm-docsis/+/678
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>
Reviewed-by: Panimalar Palaniappan <ppalaniappan.contractor@libertyglobal.com>
Reviewed-by: Mike Anderson <mike@meldnetworks.com>